move modules, create router/index
