+++ src/router/index.js
@@ -1,0 +1,5 @@
+export { default as AsyncRoute } from './AsyncRoute';
+export { default as NotFound } from './NotFound';
+export { default as Redirect } from './Redirect';
+export { default as RouteLayout } from './RouteLayout';
+export { default as SyncContainer } from './SyncContainer';
